Overhaul wrappers Locale and Holidata
