+++ bin/holidata
@@ -37,7 +37,7 @@ import sys
 
 from docopt import docopt
 
-from holidata import Holidata
+from holidata import for_locale
 
 
 def parse_locale(locale):
@@ -65,7 +65,7 @@ if __name__ == "__main__":
             # When neither '--locale' nor '--country' are given, docopt prints usage
             sys.exit(1)
 
-        print(Holidata(country=country_id, language=lang_id, year=args["--year"]).formatted_as(args["--output"]), end="")
+        print(for_locale(country_id, lang_id).holidays_of(args["--year"]).formatted_as(args["--output"]), end="")
 
     except ValueError as e:
         sys.exit(str(e))
